refactor : MoviesResponse type 응집도 강화
